transliterator/requirements.txt
pymupdf>=1.24.0
pymupdf4llm>=0.0.10
beautifulsoup4>=4.12.0
markdownify>=0.13.0
requests>=2.31.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
